fix: Improve Swarm initialization to check IP addresses
Rather than assuming the server only has one IP address, and that
everything will go ok, it checks IP addresses first. If there is one,
use it. If multiple, let the user pick.

Furthermore, the whole Swarm setup part is put in a loop, and if the
setup fails, it starts over. This allows the user to try again with
different settings, or to skip Swarm setup. Either way, it's better than
just bombing out leaving the user wondering what to do now.

Fixes #9.