feat: Add support for dnf and xcode-select when installing git
The dnf command works for RHEL, and xcode-select is for macOS. Since
macOS is a little weird about how it handles git (it has a shim for the
git binary which launches the developer tools installer), we can't just
check for the existence of git. So, for macOS only, we check the
existence of the developer tools.

What's more, once kicking off the developer tools installation, we have
to keep a loop that waits for the GUI installer to finish.