feat: Put shared installers in a "shared" directory
- Moved shared code into the "shared" directory
- Separated ghcr.io authentication logic
- Improved GitHub runner script to autodetect OS and arch
- Set up a flag "--title" where they will output their friendly title,
  which will be used by the menu
- Enhanced setup script logic to more dynamically generate the script
  menu and use the friendly "--title" flag

The enhanced setup script can now operate on the premise that all of the
scripts it will run will actually just be symbolic links to the real
scripts. So, similar to an Apache setup, we can enable/disable/reorder
what we want to implement for a given flavor by creating symbolic links
that start with a number, like "1-docker.sh".