fix: Improve podman-install-service to connect to the service
The script uses `systemctl enable "$svc_name"` to enable the newly-created service, but
since the container/pod was not started with systemctl, it is unaware of the status of the
service. This leaves systemctl unable to control the service properly in the current
session (rebooting the system fixes it since systemctl starts the service on startup).

Enhanced the script to do `systemctl enable "$svc_name" --now`, which basically tells it
to `systemctl start` the service. Unfortunately, this doesn't just attach to the running
container/pod, but instead restarts it. This is the only way to make systemctl aware of
the state of the service initially.

To enable the user to avoid the restart, added a `--no-restart` option that will work like
it did previously, but warns the user that systemctl won't work right until a system
restart.