feat: Improve GitHub Actions Runner installer to work with RHEL-9
Generally, the installer was compatible, but there needed to be some SELinux management,
and some changes for a non-native Docker environment like podman.

Additionally, we want the runner to run as a non-root user, but we do need the user to be
able to sudo some commands (deploy, docker, git). So, we add some rules to the sudoers
file so that the github user can't sudo normally but can sudo for just these commands.

Also updated to load the latest version offered by GitHub currently, and put in a check
that pings the GitHub API to see if a newer version exists than what we're offering. If
so, it asks the user if they want to try using the newer version.