fix: Abort shared setup script if any step returns errors (WIP!!)
WARNING: THIS IS A WIP. We should thoroughly check that all scripts will successfully end
with a success before considering this work done. Some may end with and if/and line that
may register as an exit with a non-zero exit code.

ORIGINAL COMMIT MESSAGE:

Sometimes the steps will have conditions that are critical and will abort. But the shared
setup script will just carry on to the next step. If any step returns a non-zero error
code, the process should abort.