feat: Improve Nginx Proxy Manager installer to secure secrets
It is common sense to have the installer tighten security on the `secrets` directory from
the beginning, instead of expecting the user to secure it manually later.

Backup processes that use rsync to pull down files from servers may encounter problems,
though, if the user doesn't have access to the files, even if they are using rsync with
sudo. So, we offer to set the `secrets` directory to a group that the real user has access
to. It's just more natural, convenient, and functional.