feat: Create and secure a `secrets` directory as part of stacks install
Even though Nginx Proxy Manager currently uses the `secrets` directory, it makes sense to
create this directory as part of the general stacks installer. Furthermore, it should be
secured by the installer so that this secure step doesn't have to be manually accomplished
later.

Backup processes that use rsync to pull down files from servers may encounter problems,
though, if the user doesn't have access to the files, even if they are using rsync with
sudo. So, we offer to set the `secrets` directory to a group that the real user has access
to. It's just more natural, convenient, and functional.