feat: Add publish script for building/publishing container images
Whereas the typical workflow will be to allow GitHub Actions to handle the building and
publishing for container images to our registry, there may be times where it is preferable
or even required that we build the images on a local workstation and publish them.

This script will help us to accomplish local workstation building and publishing when it
is needed.