feat: Improve installers to install all scripts in bin directory
Rather than explicitly installing (via symbolic links) the shell scripts in the bin
directories, we enhance the installers to loop over all the *.sh files in those
directories and install all that are found.

This will more gracefully handle all future scripts that are added to the bin directory to
automatically be included in the installation process.