feat: Improve OS selection in installer
Auto-detect OS flavor, and if it has a supported installer, ask the user if that is the
correct one to use. Otherwise, continue with a flavor list like previously done.

For macos, we just check if the arch is "Darwin". For all other Linux flavors, we inspect
the /etc/os-release file to construct a "flavor-major_ver" value that matches the way we
organize the installers in directories.