feat: Improve installer to be ephemeral
Instead of cloning the project and keeping it around, the project is adjusted to be
ephemeral, being downloaded to a temp directory just for the installation process. All
pertinent yaml or scripts are installed.

That said, a dev mode is supported, which will continue the pattern of using symlinks,
which is useful for development.

Also, improved the installer to not require running as root or with sudo. It will look at
the OS flavor and run sudo if necessary. For instance, macOS should not use sudo, but
other Linux flavors should.

As a result of the entire project not being saved on the system, installation is adjusted
to save needed scripts/configurations elsewhere:

- Stacks are saved to `/etc/stackname` (or in non-root installations like macOS, they are
  saved to `$HOME/.stackname`), like /etc/nginxproxymanager. If installing in dev mode, it
  will bring up the stack using the yaml files right in the repo.
- Reusable scripts like `deploy` and `publish` are saved in /usr/bin for Linux flavors and
  /usr/local/bin for macOS. If installing in dev mode, it will symlink the scripts.
  Otherwise, it will run `install` to copy the scripts to the bin directory.