fix: Always create "secrets" dir in nginxproxymanager installer
In dev mode, the nginxproxymanager installer could break if the "secrets" directory does
not exist. It should be created regardless of dev mode; its path will just change based on
which mode you're in.

For example, when NOT in dev mode, the secrets directory is created at one of these:
- ~/.nginxproxymanager/secrets
- /etc/nginxproxymanager/secrets

In dev mode, it will be in the project checkout:
- proj/stacks/secrets