fix: Improve build hash to be consistent between systems
The build hash was including some files that were not intended, and not sorting the files
in a consistent way, so this led to inconsistent hashes between systems.

Both `exp` and `.github` directories were allowed to be included in the hash calculation,
even though these directories are not part of the live code. Incidentally, different
locales would handle sorting `.github` differently, leading to different hashes.

So, we prune those directories from the file search, and we set `LC_ALL=C` to ensure
consistent sorting.

Additionally, we were not considering symlinks when identifying files and directories to
include in the hash calculation. This wasn't causing a problem but it could result in
unchanged hashes even when symlinks are changed or omitted. So, we improve the file search
to include symlinks in the hash calculation.